modified the way we check if account has enough funds to withdraw.
Added a do while loop to keep trying to complete the sql transaction untill successful.

Modified the BankAccount entity to not allow negative values.

fixed the wrong order of expected and actual results the assert statements,

added some code in the client so we can see the sum of the random operations.